Placing Social Media Links
The social media links were added in the navigation bar.
With the help of the HStack Chakra UI component the
social media links were placed horizontally with the proper
spacing using the "spacing" prop.

The links were added by iterating the "social" array elements
using the transformation method "map" and then returning an
anchor element (<a>) which contains a FontAwesomeIcon component.