Building The Landing Section
The Landing Section was completed by adding four components in
Chakra UI VStack component:
--> One Chakra UI Avatar component,
--> Three Chakra UI Heading components

With the aim of proper spacing to be added between the components,
the Avatar and greeting heading were wrapped into a VStack and
the remaining Heading components into another one.

The proper "spacing" prop was added to the parent VStack.

*** PS ***
  I deleted some links to my personal social media to adhere to the
  instructions
**********